backend/requirements.txt
anyio==4.6.2.post1
anytree==2.12.1
click==8.1.7
dnspython==2.7.0
exceptiongroup==1.2.2
fastapi==0.95.2
filelock==3.16.1
fsspec==2024.10.0
h11==0.14.0
idna==3.10
Jinja2==3.1.4
joblib==1.4.2
MarkupSafe==3.0.2
motor==3.1.2
mpmath==1.3.0
networkx==3.4.2
numpy==2.1.2
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
opencv-python==4.10.0.84
pillow==11.0.0
pydantic==1.10.18
pymongo==4.3.3
python-multipart==0.0.12
PyYAML==6.0.2
scikit-learn==1.5.2
scipy==1.14.1
six==1.16.0
sniffio==1.3.1
starlette==0.27.0
sympy==1.13.1
threadpoolctl==3.5.0
torch==2.5.0
torchvision==0.20.0
triton==3.1.0
typing_extensions==4.12.2
uvicorn==0.22.0